requirements.txt
psycopg2-binary
django-import-export
djangorestframework
django-cors-headers